+++ skrub/__init__.py
@@ -32,6 +32,7 @@ from ._deduplicate import compute_ngram_distance, deduplicate
 from ._drop_uninformative import DropUninformative
 from ._fuzzy_join import fuzzy_join
 from ._gap_encoder import GapEncoder
+from ._interdependence_score import interdependence_score
 from ._interpolation_joiner import InterpolationJoiner
 from ._joiner import Joiner
 from ._minhash_encoder import MinHashEncoder
@@ -103,4 +104,5 @@ __all__ = [
     "config_context",
     "ApplyToCols",
     "ApplyToFrame",
+    "interdependence_score",
 ]